Compute the index of face up card
The optional indexOfTheOneAndOnlyCard is storing data that
can be computed from the cards array.
To avoid optional getting out of sync with the data in the cards array,
change optional to a function that computes the index.

Add Array extension var to store the first element in the array if
array has only one element, otherwise nil